Make slicing and narrowing casts explicit
Make the following implicit slicing and narrowing into explicit:
```
L1TopoEvent/Root/GenericTOB.cxx:30:4: warning: slicing object from type 'JetTOB' to 'BaseTOB' discards 48 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(jet)
   ^
L1TopoEvent/Root/GenericTOB.cxx:30:4: warning: slicing object from type 'JetTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:30:4: warning: slicing object from type 'JetTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:44:4: warning: slicing object from type 'jTauTOB' to 'BaseTOB' discards 40 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(tau)
   ^
L1TopoEvent/Root/GenericTOB.cxx:44:4: warning: slicing object from type 'jTauTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:44:4: warning: slicing object from type 'jTauTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:47:12: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   , m_phi(tau.phi())
           ^
L1TopoEvent/Root/GenericTOB.cxx:56:4: warning: slicing object from type 'jLargeRJetTOB' to 'BaseTOB' discards 40 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(jet)
   ^
L1TopoEvent/Root/GenericTOB.cxx:56:4: warning: slicing object from type 'jLargeRJetTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:56:4: warning: slicing object from type 'jLargeRJetTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:59:12: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   , m_phi(jet.phi())
           ^
L1TopoEvent/Root/GenericTOB.cxx:68:4: warning: slicing object from type 'jJetTOB' to 'BaseTOB' discards 40 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(jet)
   ^
L1TopoEvent/Root/GenericTOB.cxx:68:4: warning: slicing object from type 'jJetTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:68:4: warning: slicing object from type 'jJetTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:71:12: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   , m_phi(jet.phi())
           ^
L1TopoEvent/Root/GenericTOB.cxx:80:4: warning: slicing object from type 'ClusterTOB' to 'BaseTOB' discards 44 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(cluster)
   ^
L1TopoEvent/Root/GenericTOB.cxx:80:4: warning: slicing object from type 'ClusterTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:80:4: warning: slicing object from type 'ClusterTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:92:4: warning: slicing object from type 'eEmTOB' to 'BaseTOB' discards 56 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(eem)
   ^
L1TopoEvent/Root/GenericTOB.cxx:92:4: warning: slicing object from type 'eEmTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:92:4: warning: slicing object from type 'eEmTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:95:12: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   , m_phi(eem.phi())
           ^
L1TopoEvent/Root/GenericTOB.cxx:107:4: warning: slicing object from type 'eTauTOB' to 'BaseTOB' discards 56 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(etau)
   ^
L1TopoEvent/Root/GenericTOB.cxx:107:4: warning: slicing object from type 'eTauTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:107:4: warning: slicing object from type 'eTauTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:110:12: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   , m_phi(etau.phi())
           ^
L1TopoEvent/Root/GenericTOB.cxx:122:4: warning: slicing object from type 'MuonTOB' to 'BaseTOB' discards 80 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(muon)
   ^
L1TopoEvent/Root/GenericTOB.cxx:122:4: warning: slicing object from type 'MuonTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:122:4: warning: slicing object from type 'MuonTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:139:4: warning: slicing object from type 'LateMuonTOB' to 'BaseTOB' discards 40 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(lateMuon)
   ^
L1TopoEvent/Root/GenericTOB.cxx:139:4: warning: slicing object from type 'LateMuonTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:139:4: warning: slicing object from type 'LateMuonTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:151:4: warning: slicing object from type 'MuonNextBCTOB' to 'BaseTOB' discards 40 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(muonNextBC)
   ^
L1TopoEvent/Root/GenericTOB.cxx:151:4: warning: slicing object from type 'MuonNextBCTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:151:4: warning: slicing object from type 'MuonNextBCTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:165:4: warning: slicing object from type 'MetTOB' to 'BaseTOB' discards 40 bytes of state [cppcoreguidelines-slicing]
   BaseTOB(met)
   ^
L1TopoEvent/Root/GenericTOB.cxx:165:4: warning: slicing object from type 'MetTOB' to 'BaseTOB' discards override 'print' [cppcoreguidelines-slicing]
L1TopoEvent/Root/GenericTOB.cxx:165:4: warning: slicing object from type 'MetTOB' to 'BaseTOB' discards override 'tobType' [cppcoreguidelines-slicing]
L1TopoEvent/Root/MuonTOB.cxx:26:13: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   , m_phi( sizeCheck(phi, nBitsPhi()) )
            ^
/workspaces/build/../athena/Trigger/TrigT1/L1Topo/L1TopoEvent/L1TopoEvent/JetTOB.h:36:46: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
      int energy(JetSize js ) const { return js==JS1?m_Et1:m_Et2; }
                                             ^
/workspaces/build/../athena/Trigger/TrigT1/L1Topo/L1TopoEvent/L1TopoEvent/JetTOB.h:37:42: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
      int Et(JetSize js ) const { return js==JS1?m_Et1:m_Et2; }
```

+++ Trigger/TrigT1/L1Topo/L1TopoEvent/Root/MuonTOB.cxx
@@ -23,7 +23,7 @@ TCS::MuonTOB::MuonTOB(unsigned int et, unsigned int isolation, int eta, unsigned
    , m_Et( sizeCheck(et, nBitsEt()) )
    , m_isolation( sizeCheck( isolation, nBitsIsolation()) )
    , m_eta( sizeCheck(eta, nBitsEta()) )
-   , m_phi( sizeCheck(phi, nBitsPhi()) )
+   , m_phi( static_cast<int>(sizeCheck(phi, nBitsPhi())) )
 {}
 
 // copy constructor
